requirements.txt
azure-functions
azure-identity
azure-mgmt-resourcegraph
azure-mgmt-compute
azure-data-tables
